demo_repo/brower_use2/pyproject.toml
[project]
name = "brower-use2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.12.1",
    "langchain-openai>=1.1.9",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
]
